refactor(front): extract document card component
